requirements.txt
discord.py
git+https://github.com/Rapptz/discord-ext-menus
gspread
datetime
pytz
oauth2client
jellyfish
requests
azurlane
pillow
numpy
psycopg2-binary
python-dotenv
pathlib
bs4
selenium
webdriver_manager
matplotlib